Add FastMM_ThreadContentionCount
This variable is incremented every time all the arenas for the block size are locked simultaneously and FastMM had to relinquish the thread's timeslice.  If this number is excessively high then it is an indication that the number of small, medium and/or large block arenas are insufficient for the number of application threads and should be increased.  (The CSmallBlockArenaCount, CMediumBlockArenaCount and CLargeBlockArenaCount constants.)