FastMM5 support for macOS
This version of FastMM5 includes experimental support for macOS. The ARM CPU (M1, M2, ...) architecture is weakly ordered. This could lead to really strange access violations when using threads. I added calls to MemoryBarrier to fix these. I'm not sure if I found all locations that need this. It might still be helpful for finding memory problems in the debug mode. So be careful when using this in production mode.